% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GetCovidEpi}
\alias{GetCovidEpi}
\title{GetCovidEpi}
\usage{
GetCovidEpi()
}
\description{
This function downloads the current Covid-19 US Epidemiological
data from the NTY/GitHub site and appends usefule county information
}
\examples{
GetCovidEpi()

}
\keyword{Covid19}
